+++ projects/ngx-uc/src/lib/ngx-uc-zoom-view/uc-zoom-view-image-manager.spec.ts
@@ -636,4 +636,8 @@ describe('UcZoomViewImageManager', () => {
     expect(zoomViewManager['resetZoomView']).not.toHaveBeenCalled();
   });
 
+  it('.image should return the image object of this component', () => {
+    expect(zoomViewManager.image).toBe(image);
+  });
+
 });